chore(web): extract generic time/PRNG helpers out of mocks/util.ts into lib/
Fleet parity: apparule/expendit keep generic reusable helpers under
web/src/lib/, not folded into a mock-domain module. iso()/SECOND/MINUTE/
HOUR/DAY move to lib/format.ts and hashSeed()/mulberry32()/unitFor() to
lib/random.ts — both already crossed the mocks/ boundary via
controllers/home.ts's demo chart, so they weren't mock-store-specific to
begin with. mocks/util.ts keeps only the store-bound helpers (nextId,
snapStepMs, stepLabel).
