+++ web/src/app/api/mock/v1/channels/route.ts
@@ -1,7 +1,8 @@
 import type { AlertChannel } from "@/models";
 import { jsonError, jsonOk, readJson } from "@/mocks/http";
 import { getDb } from "@/mocks/store";
-import { iso, nextId } from "@/mocks/util";
+import { iso } from "@/lib/format";
+import { nextId } from "@/mocks/util";
 
 /** GET /v1/channels — alert channels (api.md §1a; pages.md B8). */
 export async function GET() {
