+++ web/src/app/api/mock/v1/dashboards/[id]/widgets/[widgetId]/route.ts
@@ -1,7 +1,7 @@
 import type { Widget } from "@/models";
 import { jsonError, jsonOk, notFound, readJson } from "@/mocks/http";
 import { getDb } from "@/mocks/store";
-import { iso } from "@/mocks/util";
+import { iso } from "@/lib/format";
 
 export async function PATCH(
   req: Request,
