+++ web/src/app/api/mock/v1/dashboards/route.ts
@@ -1,6 +1,7 @@
 import { jsonError, jsonOk, readJson } from "@/mocks/http";
 import { getDb } from "@/mocks/store";
-import { iso, nextId } from "@/mocks/util";
+import { iso } from "@/lib/format";
+import { nextId } from "@/mocks/util";
 
 /** GET /v1/dashboards — list (org-shared + favorites, pages.md B2). */
 export async function GET() {
