+++ web/src/app/api/mock/v1/incidents/[id]/postmortem/route.ts
@@ -1,7 +1,7 @@
 import type { Postmortem, PostmortemInput } from "@/models";
 import { jsonError, jsonOk, notFound, readJson } from "@/mocks/http";
 import { getDb } from "@/mocks/store";
-import { iso } from "@/mocks/util";
+import { iso } from "@/lib/format";
 
 /**
  * B9 postmortem template (pages.md B9 "postmortem doc template on
