+++ web/src/app/api/mock/v1/incidents/route.ts
@@ -1,7 +1,8 @@
 import type { Incident, IncidentSev } from "@/models";
 import { jsonError, jsonOk, readJson } from "@/mocks/http";
 import { getDb } from "@/mocks/store";
-import { iso, nextId } from "@/mocks/util";
+import { iso } from "@/lib/format";
+import { nextId } from "@/mocks/util";
 
 /** GET /v1/incidents — list (pages.md B9). */
 export async function GET() {
