+++ web/src/app/api/mock/v1/views/route.ts
@@ -1,7 +1,8 @@
 import type { SavedView } from "@/models";
 import { jsonError, jsonOk, readJson } from "@/mocks/http";
 import { getDb } from "@/mocks/store";
-import { iso, nextId } from "@/mocks/util";
+import { iso } from "@/lib/format";
+import { nextId } from "@/mocks/util";
 
 /** GET /v1/views — saved views, optionally scoped to a surface (MI-18). */
 export async function GET(req: Request) {
