+++ web/src/mocks/seed.ts
@@ -29,16 +29,9 @@ import type {
   Trace,
   TraceSummary,
 } from "@/models";
-import {
-  DAY,
-  HOUR,
-  MINUTE,
-  SECOND,
-  iso,
-  mulberry32,
-  hashSeed,
-  nextId,
-} from "./util";
+import { DAY, HOUR, MINUTE, SECOND, iso } from "@/lib/format";
+import { hashSeed, mulberry32 } from "@/lib/random";
+import { nextId } from "./util";
 import { SERVICES, type OutageWindow } from "./series";
 import { seedSynthetics } from "./synthetics";
